+++ src/api/controllers/reservations.controllers.ts
@@ -1,9 +1,9 @@
 import { Response, Request } from 'express'
 
 import { addReservation, fetchAllReservations, fetchReservation } from '../services/reservations.service'
-import { UserInterface } from '../../types/user'
-import { INewReservation } from '../../types/reservations'
+import { UserInterface, NewReservationInterface } from '../../types/user'
 import { ReservationInstance } from '../../database/models/reservations'
+import { stripeCharge } from '../../helpers/stripe'
 
 /**
  * @export
@@ -15,15 +15,36 @@ import { ReservationInstance } from '../../database/models/reservations'
 export const createReservation = async (req: Request, res: Response): Promise<Response<any>> => {
   try {
     const {
-      body: { date, time, type, persons },
-      user: { id }
-    } = (req as unknown) as { body: INewReservation; user: UserInterface }
-    const reservation = (await addReservation(id, date, time, type, persons * 1000, persons)) as ReservationInstance
+      body: { date, time, type, stripeToken, persons },
+      user: { id, email }
+    } = (req as unknown) as { body: NewReservationInterface; user: UserInterface }
+
+    const stripeCharges = await stripeCharge(persons * 1000, 'usd', stripeToken, email)
+
+    if (stripeCharges.statusCode === 400) {
+      return res.status(400).send({
+        error: {
+          code: stripeCharges.code,
+          message: stripeCharges.message,
+          field: stripeCharges.param
+        }
+      })
+    }
+
+    const newReservation = (await addReservation(
+      id,
+      date,
+      time,
+      type,
+      persons * 1000,
+      persons,
+      stripeCharges.id
+    )) as ReservationInstance
 
     return res.status(200).send({
       status: 'success',
       data: {
-        reservation
+        reservation: newReservation
       }
     })
   } catch (error) {
