+++ src/api/middlewares/validation/reservation.validation.ts
@@ -1,7 +1,7 @@
 import { Request, Response, NextFunction } from 'express'
 
 import { validationHandler } from './validationHandler'
-import { createReservationSchema, reservationChargeSchema } from './schemas/reservations.schema'
+import { createReservationSchema } from './schemas/reservation.schema'
 
 /**
  * @export
@@ -13,16 +13,4 @@ import { createReservationSchema, reservationChargeSchema } from './schemas/rese
  */
 export const createReservationValidator = (req: Request, res: Response, next: NextFunction) => {
   validationHandler(req, res, next, createReservationSchema)
-}
-
-/**
- * @export
- * @function reservationChargeValidator
- * @param {Object} req - request received
- * @param {Object} res - response object
- * @param {Object} next - next object
- * @returns {Object} next object
- */
-export const reservationChargeValidator = (req: Request, res: Response, next: NextFunction) => {
-  validationHandler(req, res, next, reservationChargeSchema)
 }
