+++ src/api/routes/api/reservations.router.ts
@@ -2,9 +2,8 @@ import { Router } from 'express'
 import passport from 'passport'
 
 import { authInterceptor } from '../../middlewares/validation/validationHandler'
-import { createReservationValidator, reservationChargeValidator } from '../../middlewares/validation/reservations.validation'
+import { createReservationValidator } from '../../middlewares/validation/reservation.validation'
 import { createReservation, getAllReservations, getReservation } from '../../controllers/reservations.controllers'
-import { createReservationCharge } from '../../controllers/charges.controllers'
 
 const reservationsRouter = Router()
 
@@ -17,12 +16,5 @@ reservationsRouter.post(
   createReservation
 )
 reservationsRouter.get('/:id', passport.authenticate('jwt', { session: false }), authInterceptor, getReservation)
-reservationsRouter.post(
-  '/:id/charges',
-  passport.authenticate('jwt', { session: false }),
-  authInterceptor,
-  reservationChargeValidator,
-  createReservationCharge
-)
 
 export default reservationsRouter
