+++ src/api/services/reservations.service.ts
@@ -10,6 +10,7 @@ import { ReservationInstance } from '../../database/models/reservations'
  * @param {Integer} price - reservation price
  * @param {Integer} persons - number of people
  * @param {String} type - type of menu
+ * @param {String} stripeId - stripeId
  * @returns {Object} object
  */
 export const addReservation = (
@@ -18,10 +19,11 @@ export const addReservation = (
   time: string,
   type: string,
   price: number,
-  persons: number
+  persons: number,
+  stripeId: string
 ): Promise<ReservationInstance> => {
   const newDate: any = new Date(date)
-  const reservation = db.Reservation.create({ userId, date: newDate, time, type, price, persons })
+  const reservation = db.Reservation.create({ userId, date: newDate, time, type, price, persons, stripeId })
   return reservation
 }
