+++ src/tests/mocks/reservations.mock.ts
@@ -2,11 +2,13 @@ export const newReservation = {
   date: '8/31/2020',
   time: '23:00',
   persons: 2,
-  type: 'dinner'
+  type: 'dinner',
+  stripeToken: 'tok_visa'
 }
 
 export const incorrectNewReservation = {
   date: '',
   persons: 2,
-  type: 'dinner'
+  type: 'dinner',
+  stripeToken: 'tok_visa'
 }
